refactor(config): move generate_username out of the OIDC backend (refactor-survival demo)
#957 refactor-survival demonstration: behavior tests survive a real refactor that
mock-coupled tests would not.

Move the standalone Django-username validator generate_username (and its
DJANGO_USERNAME_* constants) from config/oidc.py into a dedicated config/username.py
and update the single topology reference OIDC_USERNAME_ALGO to
'config.username.generate_username'. The validator never belonged in the OIDC
backend module.

The tests/mission_control/test_oidc.py ShifterOIDCBackend behavior tests exercise
generate_username only through the real OIDC create_user/update_user login flow
(via OIDC_USERNAME_ALGO), never naming the function's location, and pass UNCHANGED
across the move. Only the function's own direct unit tests (TestGenerateUsername)
updated their import to follow it. A mock-coupled test that had patched
config.oidc.generate_username would have broken; the behavior tests did not.

No behavior change.
